+++ Sources/ShortcutRecorder/SRShortcut.m
@@ -3,8 +3,6 @@
 //  CC BY 4.0
 //
 
-#import <os/log.h>
-
 #import "ShortcutRecorder/SRCommon.h"
 #import "ShortcutRecorder/SRKeyCodeTransformer.h"
 #import "ShortcutRecorder/SRShortcutFormatter.h"
@@ -18,6 +16,9 @@ SRShortcutKey const SRShortcutKeyKeyCode = @"keyCode";
 SRShortcutKey const SRShortcutKeyModifierFlags = @"modifierFlags";
 SRShortcutKey const SRShortcutKeyCharacters = @"characters";
 SRShortcutKey const SRShortcutKeyCharactersIgnoringModifiers = @"charactersIgnoringModifiers";
+
+
+static os_log_t _Log;
 
 
 @implementation SRShortcut
@@ -43,7 +44,7 @@ SRShortcutKey const SRShortcutKeyCharactersIgnoringModifiers = @"charactersIgnor
     __auto_type eventType = aKeyboardEvent.type;
     if (((1 << eventType) & (NSEventMaskKeyDown | NSEventMaskKeyUp | NSEventMaskFlagsChanged)) == 0)
     {
-        os_log_error(OS_LOG_DEFAULT, "#Error aKeyboardEvent must be either NSEventTypeKeyUp, NSEventTypeKeyDown or NSEventTypeFlagsChanged, but got %lu", aKeyboardEvent.type);
+        SRLogError(_Log, "aKeyboardEvent must be NSEventTypeKeyDown (10), NSEventTypeKeyUp (11) or NSEventTypeFlagsChanged (12), got %lu instead", aKeyboardEvent.type);
         return nil;
     }
 
@@ -80,7 +81,7 @@ SRShortcutKey const SRShortcutKeyCharactersIgnoringModifiers = @"charactersIgnor
                 if (!NSThread.isMainThread)
                 {
                     NSParameterAssert(NO);
-                    os_log_error(OS_LOG_DEFAULT, "#Error #Developer AppKit failed to extract characters because it is used in a non-main thread, see SRShortcut/shortcutWithEvent:ignoringCharacters:");
+                    SRLogError(_Log, "AppKit failed to extract characters because it is used in a non-main thread, see SRShortcut/shortcutWithEvent:ignoringCharacters:");
                 }
                 else
                     @throw;
@@ -166,7 +167,7 @@ SRShortcutKey const SRShortcutKeyCharactersIgnoringModifiers = @"charactersIgnor
       charactersIgnoringModifiers:nil];
 }
 
-+ (nullable instancetype)shortcutWithKeyBinding:(NSString *)aKeyBinding
++ (instancetype)shortcutWithKeyBinding:(NSString *)aKeyBinding
 {
     return [SRKeyBindingTransformer.sharedTransformer transformedValue:aKeyBinding];
 }
@@ -374,7 +375,7 @@ SRShortcutKey const SRShortcutKeyCharactersIgnoringModifiers = @"charactersIgnor
 
 #pragma mark Subscript
 
-- (nullable id)objectForKeyedSubscript:(SRShortcutKey)aKey
+- (id)objectForKeyedSubscript:(SRShortcutKey)aKey
 {
     if ([aKey isEqualToString:SRShortcutKeyKeyCode])
         return @(self.keyCode);
@@ -425,13 +426,19 @@ SRShortcutKey const SRShortcutKeyCharactersIgnoringModifiers = @"charactersIgnor
 
 #pragma mark NSObject
 
-#ifndef __clang_analyzer__
++ (void)initialize
+{
+    static dispatch_once_t OnceToken;
+    dispatch_once(&OnceToken, ^{
+        _Log = os_log_create(SRLogSubsystem.UTF8String, SRLogCategoryKeyBindingTransformer.UTF8String);
+    });
+}
+
 + (instancetype)new
 {
     [self doesNotRecognizeSelector:_cmd];
     return nil;
 }
-#endif
 
 - (instancetype)init
 {
@@ -472,7 +479,7 @@ SRShortcutKey const SRShortcutKeyCharactersIgnoringModifiers = @"charactersIgnor
 - (UInt32)carbonKeyCode
 {
     if (self.keyCode == SRKeyCodeNone)
-        os_log_error(OS_LOG_DEFAULT, "#Critical SRKeyCodeNone has no representation in Carbon");
+        SRLogError(_Log, "SRKeyCodeNone has no representation in Carbon");
 
     return self.keyCode;
 }